refactoring, code cleanup, preparing spring-batch
