+++ src/main/java/org/woehlke/twitterwall/oodm/service/UserService.java
@@ -9,16 +9,10 @@ import java.util.List;
 /**
  * Created by tw on 11.06.17.
  */
-public interface UserService extends DomainService {
+public interface UserService extends DomainService<User> {
 
     User findByIdTwitter(long idTwitter);
-
-    User persist(User user);
-
-    User update(User user);
-
-    List<User> getAll();
-
+    
     User findByScreenName(String screenName);
 
     List<User> getTweetingUsers();
@@ -29,13 +23,5 @@ public interface UserService extends DomainService {
 
     List<Long> getAllTwitterIds();
 
-    User storeUser(User user);
-
     User storeUserProcess(User user);
-
-    User transform(TwitterProfile twitterProfile);
-
-    User getEntitiesForUrlDescription(User user);
-
-    User storeUserProfile(TwitterProfile userProfile);
 }
