+++ src/main/java/org/woehlke/twitterwall/oodm/service/entities/UrlCacheService.java
@@ -7,7 +7,7 @@ import org.woehlke.twitterwall.oodm.service.common.DomainService;
 /**
  * Created by tw on 23.06.17.
  */
-public interface UrlCacheService extends DomainService {
+public interface UrlCacheService extends DomainService<UrlCache> {
 
     UrlCache store(UrlCache urlCache);
