+++ src/main/java/org/woehlke/twitterwall/oodm/service/entities/UrlService.java
@@ -11,26 +11,14 @@ import java.util.Set;
 /**
  * Created by tw on 12.06.17.
  */
-public interface UrlService extends DomainService {
-
-    Url create(Url url);
-
-    Url update(Url url);
+public interface UrlService extends DomainService<Url> {
 
     Url findByDisplayExpandedUrl(String display, String expanded, String url);
 
     Url findByUrl(String url);
 
-    Url store(Url url);
-
-    Set<Url> transformUrls(List<UrlEntity> urls);
-
-    Url fetchTransientUrl(String urlSrc);
-
     Url getPersistentUrlFor(String url);
 
     List<Url> getTestData();
-
-    Set<Url> getUrlsFor(User user);
-
+    
 }
