+++ src/main/java/org/woehlke/twitterwall/process/backend/service/entities/MediaTransformServiceImpl.java
@@ -1,0 +1,29 @@
+package org.woehlke.twitterwall.process.backend.service.entities;
+
+import org.springframework.social.twitter.api.MediaEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+import org.woehlke.twitterwall.oodm.entities.entities.Media;
+
+/**
+ * Created by tw on 28.06.17.
+ */
+@Service
+@Transactional(propagation = Propagation.REQUIRED, readOnly = true)
+public class MediaTransformServiceImpl implements MediaTransformService {
+    
+    @Override
+    public Media transform(MediaEntity medium) {
+        long idTwitter = medium.getId();
+        String mediaHttp = medium.getMediaUrl();
+        String mediaHttps = medium.getMediaSecureUrl();
+        String url = medium.getUrl();
+        String display = medium.getDisplayUrl();
+        String expanded = medium.getExpandedUrl();
+        String type = medium.getType();
+        int[] indices = medium.getIndices();
+        Media myMediaEntity = new Media(idTwitter, mediaHttp, mediaHttps, url, display, expanded, type, indices);
+        return myMediaEntity;
+    }
+}
